% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config.R
\name{FishlifeLHTNameSpace}
\alias{FishlifeLHTNameSpace}
\title{Fishlife Context class}
\description{
Class that allows the transition between user-defined LHT names and those specifically used by Fishlife
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{updated_prefix}}{string added to each obtained LHT from FishLife}

\item{\code{lht_names}}{User-defined list of LHT names associated with the expected FishLife names}

\item{\code{transform_function}}{Forward function conversion required for each user-defined parameter.}

\item{\code{backtransform_function}}{Backward function conversion required for each obtained LHT value.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-FishlifeLHTNameSpace-new}{\code{FishlifeLHTNameSpace$new()}}
\item \href{#method-FishlifeLHTNameSpace-clone}{\code{FishlifeLHTNameSpace$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-FishlifeLHTNameSpace-new"></a>}}
\if{latex}{\out{\hypertarget{method-FishlifeLHTNameSpace-new}{}}}
\subsection{Method \code{new()}}{
Initialise Fishlife Context class
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FishlifeLHTNameSpace$new(
  updated_prefix = "updated",
  lht_names = list(Linf = "log(length_infinity)", Winf = "log(weight_infinity)", K =
    "log(growth_coefficient)", M = "log(natural_mortality)", L50 =
    "log(length_maturity)", Amax = "log(age_max)", Amat = "log(age_maturity)",
    Temperature = "temperature"),
  transform_function = list(`log(length_infinity)` = log, `log(weight_infinity)` = log,
    `log(growth_coefficient)` = log, `log(natural_mortality)` = log,
    `log(length_maturity)` = log, `log(age_max)` = log, `log(age_maturity)` = log,
    temperature = base::identity),
  backtransform_function = list(`log(length_infinity)` = exp, `log(weight_infinity)` =
    exp, `log(growth_coefficient)` = exp, `log(natural_mortality)` = exp,
    `log(length_maturity)` = exp, `log(age_max)` = exp, `log(age_maturity)` = exp,
    temperature = base::identity)
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-FishlifeLHTNameSpace-clone"></a>}}
\if{latex}{\out{\hypertarget{method-FishlifeLHTNameSpace-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FishlifeLHTNameSpace$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
