% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{MixinUtilities}
\alias{MixinUtilities}
\title{Mixin Utility class}
\description{
class that holds a few util methods used across all classes in the this R package
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-MixinUtilities-create_empty_dataframe}{\code{MixinUtilities$create_empty_dataframe()}}
\item \href{#method-MixinUtilities-list_of_vectors_to_dataframe}{\code{MixinUtilities$list_of_vectors_to_dataframe()}}
\item \href{#method-MixinUtilities-apply_func_to_df}{\code{MixinUtilities$apply_func_to_df()}}
\item \href{#method-MixinUtilities-clone}{\code{MixinUtilities$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-MixinUtilities-create_empty_dataframe"></a>}}
\if{latex}{\out{\hypertarget{method-MixinUtilities-create_empty_dataframe}{}}}
\subsection{Method \code{create_empty_dataframe()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MixinUtilities$create_empty_dataframe(col_names)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-MixinUtilities-list_of_vectors_to_dataframe"></a>}}
\if{latex}{\out{\hypertarget{method-MixinUtilities-list_of_vectors_to_dataframe}{}}}
\subsection{Method \code{list_of_vectors_to_dataframe()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MixinUtilities$list_of_vectors_to_dataframe(lht_list)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-MixinUtilities-apply_func_to_df"></a>}}
\if{latex}{\out{\hypertarget{method-MixinUtilities-apply_func_to_df}{}}}
\subsection{Method \code{apply_func_to_df()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MixinUtilities$apply_func_to_df(data, func_space)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-MixinUtilities-clone"></a>}}
\if{latex}{\out{\hypertarget{method-MixinUtilities-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MixinUtilities$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
