% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predicted_lht_generator.R
\name{PredictedLHTGenerator}
\alias{PredictedLHTGenerator}
\title{Convert a dataframe into a list of rows removing those columns per species that contain NA
Given a list of LHTs it renames each variable into the expected FishLife names
Build a list of predicting parameters for each species
Generate the predicted Fishlife matrices (LHTs and LHT covariances)}
\description{
Convert a dataframe into a list of rows removing those columns per species that contain NA
Given a list of LHTs it renames each variable into the expected FishLife names
Build a list of predicting parameters for each species
Generate the predicted Fishlife matrices (LHTs and LHT covariances)

Convert a dataframe into a list of rows removing those columns per species that contain NA
Given a list of LHTs it renames each variable into the expected FishLife names
Build a list of predicting parameters for each species
Generate the predicted Fishlife matrices (LHTs and LHT covariances)
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-PredictedLHTGenerator-new}{\code{PredictedLHTGenerator$new()}}
\item \href{#method-PredictedLHTGenerator-predict}{\code{PredictedLHTGenerator$predict()}}
\item \href{#method-PredictedLHTGenerator-clone}{\code{PredictedLHTGenerator$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PredictedLHTGenerator-new"></a>}}
\if{latex}{\out{\hypertarget{method-PredictedLHTGenerator-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PredictedLHTGenerator$new(
  master_db,
  lht_predicting_matrix,
  func_conversion_matrix,
  lht_names
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PredictedLHTGenerator-predict"></a>}}
\if{latex}{\out{\hypertarget{method-PredictedLHTGenerator-predict}{}}}
\subsection{Method \code{predict()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PredictedLHTGenerator$predict()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PredictedLHTGenerator-clone"></a>}}
\if{latex}{\out{\hypertarget{method-PredictedLHTGenerator-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PredictedLHTGenerator$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
