% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predicted_lht_picker.R
\name{PredictedLHTPicker}
\alias{PredictedLHTPicker}
\title{Multiple Taxon Collector class}
\value{
dataframe that keeps the LHT values per taxon

subset matrix of LHT values
}
\description{
Fetch, transform and build the predicted LHT dataframe for multiple species obtained from FishLife
}
\section{Super class}{
\code{\link[LHTpicker:MixinUtilities]{LHTpicker::MixinUtilities}} -> \code{PredictedLHTPicker}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-PredictedLHTPicker-new}{\code{PredictedLHTPicker$new()}}
\item \href{#method-PredictedLHTPicker-pick_and_backtransform}{\code{PredictedLHTPicker$pick_and_backtransform()}}
\item \href{#method-PredictedLHTPicker-clone}{\code{PredictedLHTPicker$clone()}}
}
}
\if{html}{\out{
<details open><summary>Inherited methods</summary>
<ul>
<li><span class="pkg-link" data-pkg="LHTpicker" data-topic="MixinUtilities" data-id="apply_func_to_df"><a href='../../LHTpicker/html/MixinUtilities.html#method-MixinUtilities-apply_func_to_df'><code>LHTpicker::MixinUtilities$apply_func_to_df()</code></a></span></li>
<li><span class="pkg-link" data-pkg="LHTpicker" data-topic="MixinUtilities" data-id="create_empty_dataframe"><a href='../../LHTpicker/html/MixinUtilities.html#method-MixinUtilities-create_empty_dataframe'><code>LHTpicker::MixinUtilities$create_empty_dataframe()</code></a></span></li>
<li><span class="pkg-link" data-pkg="LHTpicker" data-topic="MixinUtilities" data-id="list_of_vectors_to_dataframe"><a href='../../LHTpicker/html/MixinUtilities.html#method-MixinUtilities-list_of_vectors_to_dataframe'><code>LHTpicker::MixinUtilities$list_of_vectors_to_dataframe()</code></a></span></li>
</ul>
</details>
}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PredictedLHTPicker-new"></a>}}
\if{latex}{\out{\hypertarget{method-PredictedLHTPicker-new}{}}}
\subsection{Method \code{new()}}{
Initialise the PredictedLHTPicker
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PredictedLHTPicker$new(
  master_db,
  lht_names,
  backtransform_function_list,
  wanted_lht_df
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{master_db}}{Fishlife database}

\item{\code{lht_names}}{list of user-defined LHT names associated with their FishLife's counterparts}

\item{\code{backtransform_function_list}}{list of backward-transformation functions to be applied on obtained LHT from FishLife}

\item{\code{wanted_lht_df}}{dataframe holding the wanted taxa details as rows and LHT as columns}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PredictedLHTPicker-pick_and_backtransform"></a>}}
\if{latex}{\out{\hypertarget{method-PredictedLHTPicker-pick_and_backtransform}{}}}
\subsection{Method \code{pick_and_backtransform()}}{
Pick and backtransform all LHT values per taxon that have beed obtained from Fishlife
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PredictedLHTPicker$pick_and_backtransform()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PredictedLHTPicker-clone"></a>}}
\if{latex}{\out{\hypertarget{method-PredictedLHTPicker-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PredictedLHTPicker$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
