% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/taxa_lht_collector.R
\name{TaxaLHTCollector}
\alias{TaxaLHTCollector}
\title{Multiple Taxon Collector class}
\value{
dataframe that keeps the LHT values per taxon

subset matrix of LHT values
}
\description{
Class that collect and transform predicted values of LHTs for multiple species obtained from FishLife
}
\section{Super class}{
\code{\link[LHTpicker:MixinUtilities]{LHTpicker::MixinUtilities}} -> \code{TaxaLHTCollector}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-TaxaLHTCollector-new}{\code{TaxaLHTCollector$new()}}
\item \href{#method-TaxaLHTCollector-collect_and_backtransform}{\code{TaxaLHTCollector$collect_and_backtransform()}}
\item \href{#method-TaxaLHTCollector-clone}{\code{TaxaLHTCollector$clone()}}
}
}
\if{html}{\out{
<details open><summary>Inherited methods</summary>
<ul>
<li><span class="pkg-link" data-pkg="LHTpicker" data-topic="MixinUtilities" data-id="apply_func_to_df"><a href='../../LHTpicker/html/MixinUtilities.html#method-MixinUtilities-apply_func_to_df'><code>LHTpicker::MixinUtilities$apply_func_to_df()</code></a></span></li>
<li><span class="pkg-link" data-pkg="LHTpicker" data-topic="MixinUtilities" data-id="create_empty_dataframe"><a href='../../LHTpicker/html/MixinUtilities.html#method-MixinUtilities-create_empty_dataframe'><code>LHTpicker::MixinUtilities$create_empty_dataframe()</code></a></span></li>
<li><span class="pkg-link" data-pkg="LHTpicker" data-topic="MixinUtilities" data-id="list_of_vectors_to_dataframe"><a href='../../LHTpicker/html/MixinUtilities.html#method-MixinUtilities-list_of_vectors_to_dataframe'><code>LHTpicker::MixinUtilities$list_of_vectors_to_dataframe()</code></a></span></li>
</ul>
</details>
}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TaxaLHTCollector-new"></a>}}
\if{latex}{\out{\hypertarget{method-TaxaLHTCollector-new}{}}}
\subsection{Method \code{new()}}{
Initialise the TaxaLHTCollector
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TaxaLHTCollector$new(
  master_db,
  lht_names,
  backtransform_function_list,
  wanted_lht_df
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{master_db}}{Fishlife database}

\item{\code{lht_names}}{list of user-defined LHT names associated with their FishLife's counterparts}

\item{\code{backtransform_function_list}}{list of backward-transformation functions to be applied on obtained LHT from FishLife}

\item{\code{wanted_lht_df}}{dataframe holding the wanted taxa details as rows and LHT as columns}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TaxaLHTCollector-collect_and_backtransform"></a>}}
\if{latex}{\out{\hypertarget{method-TaxaLHTCollector-collect_and_backtransform}{}}}
\subsection{Method \code{collect_and_backtransform()}}{
Collect and backtransform all LHT values per taxon that have beed obtained from Fishlife
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TaxaLHTCollector$collect_and_backtransform()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TaxaLHTCollector-clone"></a>}}
\if{latex}{\out{\hypertarget{method-TaxaLHTCollector-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TaxaLHTCollector$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
