% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/taxa_updated_lht_builder.R
\name{TaxaUpdatedLHTBuilder}
\alias{TaxaUpdatedLHTBuilder}
\title{Multiple Taxon Update Extractor class}
\value{
dataframe that keeps the original and updated LHT values per taxon

updated dataframe of results with both the original and updated LHT values
}
\description{
Class that extracts and transform predicted values of LHTs for multiple species obtained from FishLife
}
\section{Super class}{
\code{\link[LHTpicker:MixinUtilities]{LHTpicker::MixinUtilities}} -> \code{TaxaUpdatedLHTBuilder}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-TaxaUpdatedLHTBuilder-new}{\code{TaxaUpdatedLHTBuilder$new()}}
\item \href{#method-TaxaUpdatedLHTBuilder-extract_and_backtransform}{\code{TaxaUpdatedLHTBuilder$extract_and_backtransform()}}
\item \href{#method-TaxaUpdatedLHTBuilder-clone}{\code{TaxaUpdatedLHTBuilder$clone()}}
}
}
\if{html}{\out{
<details open><summary>Inherited methods</summary>
<ul>
<li><span class="pkg-link" data-pkg="LHTpicker" data-topic="MixinUtilities" data-id="apply_func_to_df"><a href='../../LHTpicker/html/MixinUtilities.html#method-MixinUtilities-apply_func_to_df'><code>LHTpicker::MixinUtilities$apply_func_to_df()</code></a></span></li>
<li><span class="pkg-link" data-pkg="LHTpicker" data-topic="MixinUtilities" data-id="create_empty_dataframe"><a href='../../LHTpicker/html/MixinUtilities.html#method-MixinUtilities-create_empty_dataframe'><code>LHTpicker::MixinUtilities$create_empty_dataframe()</code></a></span></li>
<li><span class="pkg-link" data-pkg="LHTpicker" data-topic="MixinUtilities" data-id="list_of_vectors_to_dataframe"><a href='../../LHTpicker/html/MixinUtilities.html#method-MixinUtilities-list_of_vectors_to_dataframe'><code>LHTpicker::MixinUtilities$list_of_vectors_to_dataframe()</code></a></span></li>
</ul>
</details>
}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TaxaUpdatedLHTBuilder-new"></a>}}
\if{latex}{\out{\hypertarget{method-TaxaUpdatedLHTBuilder-new}{}}}
\subsection{Method \code{new()}}{
Initialise the TaxaPredictionExtractor
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TaxaUpdatedLHTBuilder$new(
  update_prefix,
  lht_names,
  backtransform_function_list,
  predicting_lht_df,
  updated_lht_list
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{update_prefix}}{string to be added as prefix to the column names keeping the obtained new LHT values}

\item{\code{lht_names}}{list of user-defined LHT names associated with their FishLife's counterparts}

\item{\code{backtransform_function_list}}{list of backward-transformation functions to be applied on obtained LHT from FishLife}

\item{\code{predicting_lht_df}}{dataframe keeping the original inputted LHTs per taxon}

\item{\code{updated_lht_list}}{list of updated LHTs per taxon obtained from Fishlife}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TaxaUpdatedLHTBuilder-extract_and_backtransform"></a>}}
\if{latex}{\out{\hypertarget{method-TaxaUpdatedLHTBuilder-extract_and_backtransform}{}}}
\subsection{Method \code{extract_and_backtransform()}}{
Extract and backtransform all updated LHT values per taxon that have beed obtained from Fishlife
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TaxaUpdatedLHTBuilder$extract_and_backtransform()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TaxaUpdatedLHTBuilder-clone"></a>}}
\if{latex}{\out{\hypertarget{method-TaxaUpdatedLHTBuilder-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TaxaUpdatedLHTBuilder$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
