% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/taxa_updated_lht_getter.R
\name{TaxaUpdatedLHTGetter}
\alias{TaxaUpdatedLHTGetter}
\title{Multiple Updated LHT Generator class}
\value{
list of LHT matrices, where each name in the list corresponds a single taxon

a unnamed list of LHT values per taxon

renamed list of lht which names conform to Fishlife's expected patterns

list of LHT values (another list) for each taxon
}
\description{
Allows the streamline-generation of updated LHTs given new data.
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-TaxaUpdatedLHTGetter-new}{\code{TaxaUpdatedLHTGetter$new()}}
\item \href{#method-TaxaUpdatedLHTGetter-update}{\code{TaxaUpdatedLHTGetter$update()}}
\item \href{#method-TaxaUpdatedLHTGetter-clone}{\code{TaxaUpdatedLHTGetter$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TaxaUpdatedLHTGetter-new"></a>}}
\if{latex}{\out{\hypertarget{method-TaxaUpdatedLHTGetter-new}{}}}
\subsection{Method \code{new()}}{
Initialise the Updated LHT Generator class
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TaxaUpdatedLHTGetter$new(
  master_db,
  predicting_lht_df,
  transform_function_list,
  lht_names
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{master_db}}{Fishlife database}

\item{\code{predicting_lht_df}}{dataframe providing the taxon details and their predicting LHT values}

\item{\code{transform_function_list}}{list of transforming function which names must conform to FishLife's expectations}

\item{\code{lht_names}}{list of user-defined LHT names associated with their FishLife's counterparts}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TaxaUpdatedLHTGetter-update"></a>}}
\if{latex}{\out{\hypertarget{method-TaxaUpdatedLHTGetter-update}{}}}
\subsection{Method \code{update()}}{
Update the given LHTs per taxon according to the new predicted value yielded by FishLfe
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TaxaUpdatedLHTGetter$update()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TaxaUpdatedLHTGetter-clone"></a>}}
\if{latex}{\out{\hypertarget{method-TaxaUpdatedLHTGetter-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TaxaUpdatedLHTGetter$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
