% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/taxon_extractor.R
\name{TaxonExtractor}
\alias{TaxonExtractor}
\title{Single Taxon Extractor class}
\value{
A list with three elements:
   a) The database unique identifier for the sought taxon
   b) The predicted and stored LHTs for such taxon
   c) The convariance matrix for all parameters associated to the sought taxon
}
\description{
Class that abstracts away the underlying FishLife database structure and extract the details of a given taxon
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-TaxonExtractor-new}{\code{TaxonExtractor$new()}}
\item \href{#method-TaxonExtractor-extract}{\code{TaxonExtractor$extract()}}
\item \href{#method-TaxonExtractor-clone}{\code{TaxonExtractor$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TaxonExtractor-new"></a>}}
\if{latex}{\out{\hypertarget{method-TaxonExtractor-new}{}}}
\subsection{Method \code{new()}}{
Initialise Taxon Extractor class
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TaxonExtractor$new(master_db, taxon)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{master_db}}{Fishlife database}

\item{\code{taxon}}{string holding the taxon name}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TaxonExtractor-extract"></a>}}
\if{latex}{\out{\hypertarget{method-TaxonExtractor-extract}{}}}
\subsection{Method \code{extract()}}{
Extract the sought taxon details
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TaxonExtractor$extract()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TaxonExtractor-clone"></a>}}
\if{latex}{\out{\hypertarget{method-TaxonExtractor-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TaxonExtractor$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
