% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/taxon_lht_grabber.R
\name{TaxonLHTGrabber}
\alias{TaxonLHTGrabber}
\title{Single Taxon LHT GRABBER class}
\value{
A list with three elements:
   a) The database unique identifier for the sought taxon
   b) The predicted and stored LHTs for such taxon
   c) The convariance matrix for all parameters associated to the sought taxon
}
\description{
Class that abstracts away the underlying FishLife dynamics and grab the details of a given taxon
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-TaxonLHTGrabber-new}{\code{TaxonLHTGrabber$new()}}
\item \href{#method-TaxonLHTGrabber-extract}{\code{TaxonLHTGrabber$extract()}}
\item \href{#method-TaxonLHTGrabber-clone}{\code{TaxonLHTGrabber$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TaxonLHTGrabber-new"></a>}}
\if{latex}{\out{\hypertarget{method-TaxonLHTGrabber-new}{}}}
\subsection{Method \code{new()}}{
Initialise Taxon LHT Grabber class
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TaxonLHTGrabber$new(master_db, taxon)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{master_db}}{Fishlife database}

\item{\code{taxon}}{string holding the taxon name}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TaxonLHTGrabber-extract"></a>}}
\if{latex}{\out{\hypertarget{method-TaxonLHTGrabber-extract}{}}}
\subsection{Method \code{extract()}}{
Fetch the sought taxon details
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TaxonLHTGrabber$extract()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TaxonLHTGrabber-clone"></a>}}
\if{latex}{\out{\hypertarget{method-TaxonLHTGrabber-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TaxonLHTGrabber$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
