% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/taxon_lht_predictor.R
\name{TaxonLHTPredictor}
\alias{TaxonLHTPredictor}
\title{Single Taxon LHT Predictor class}
\value{
matrix-form LHT values to be passed on to FishLife
}
\description{
Class that given some LHTs will fetch their predicted version according to FishLife covariance
}
\section{Super class}{
\code{LHTPicker::MixinUtilities} -> \code{TaxonLHTPredictor}
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{common_columns_ds}}{common columns across the underlying data structure of FishLife. Taxa's LHT
could end up having additional fields that the overall covariance matrix does not support.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-TaxonLHTPredictor-new}{\code{TaxonLHTPredictor$new()}}
\item \href{#method-TaxonLHTPredictor-generate_new_lht_matrix}{\code{TaxonLHTPredictor$generate_new_lht_matrix()}}
\item \href{#method-TaxonLHTPredictor-predict}{\code{TaxonLHTPredictor$predict()}}
\item \href{#method-TaxonLHTPredictor-clone}{\code{TaxonLHTPredictor$clone()}}
}
}
\if{html}{\out{
<details open><summary>Inherited methods</summary>
<ul>
<li><span class="pkg-link" data-pkg="LHTPicker" data-topic="MixinUtilities" data-id="apply_func_to_df"><a href='../../LHTPicker/html/MixinUtilities.html#method-MixinUtilities-apply_func_to_df'><code>LHTPicker::MixinUtilities$apply_func_to_df()</code></a></span></li>
<li><span class="pkg-link" data-pkg="LHTPicker" data-topic="MixinUtilities" data-id="create_empty_dataframe"><a href='../../LHTPicker/html/MixinUtilities.html#method-MixinUtilities-create_empty_dataframe'><code>LHTPicker::MixinUtilities$create_empty_dataframe()</code></a></span></li>
<li><span class="pkg-link" data-pkg="LHTPicker" data-topic="MixinUtilities" data-id="list_of_vectors_to_dataframe"><a href='../../LHTPicker/html/MixinUtilities.html#method-MixinUtilities-list_of_vectors_to_dataframe'><code>LHTPicker::MixinUtilities$list_of_vectors_to_dataframe()</code></a></span></li>
</ul>
</details>
}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TaxonLHTPredictor-new"></a>}}
\if{latex}{\out{\hypertarget{method-TaxonLHTPredictor-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TaxonLHTPredictor$new(
  master_db,
  estimated_lhts,
  estimated_lht_cov,
  new_lhts,
  func_domains
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{master_db}}{Fishlife database}

\item{\code{estimated_lhts}}{taxon's LHT numeric vector as fetched from Fishlife.}

\item{\code{new_lhts}}{predicting LHT list which names must conform to FishLife's expectations}

\item{\code{func_domains}}{list of transforming function which names must conform to FishLife's expectations}

\item{\code{estimated_lht_conv}}{taxon's covariance matrix as fetched from Fishlife}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TaxonLHTPredictor-generate_new_lht_matrix"></a>}}
\if{latex}{\out{\hypertarget{method-TaxonLHTPredictor-generate_new_lht_matrix}{}}}
\subsection{Method \code{generate_new_lht_matrix()}}{
Generate the new LHT matrix in shape and mathematical domain expected by Fishlife
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TaxonLHTPredictor$generate_new_lht_matrix()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TaxonLHTPredictor-predict"></a>}}
\if{latex}{\out{\hypertarget{method-TaxonLHTPredictor-predict}{}}}
\subsection{Method \code{predict()}}{
Generate a matrix with the new predicted LHTs given some initial values, both in log and log-converted space
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TaxonLHTPredictor$predict()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TaxonLHTPredictor-clone"></a>}}
\if{latex}{\out{\hypertarget{method-TaxonLHTPredictor-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TaxonLHTPredictor$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
