% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/taxon_predicted_lht_getter.R
\name{TaxonPredictedLHTGetter}
\alias{TaxonPredictedLHTGetter}
\title{Single Taxon LHT GRABBER class}
\value{
A list with three elements:
   a) The database unique identifier for the sought taxon
   b) The predicted and stored LHTs for such taxon
   c) The convariance matrix for all parameters associated to the sought taxon
}
\description{
Abstracts away the underlying FishLife dynamics and grab the predicted LHT details of a given taxon
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-TaxonPredictedLHTGetter-new}{\code{TaxonPredictedLHTGetter$new()}}
\item \href{#method-TaxonPredictedLHTGetter-extract}{\code{TaxonPredictedLHTGetter$extract()}}
\item \href{#method-TaxonPredictedLHTGetter-clone}{\code{TaxonPredictedLHTGetter$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TaxonPredictedLHTGetter-new"></a>}}
\if{latex}{\out{\hypertarget{method-TaxonPredictedLHTGetter-new}{}}}
\subsection{Method \code{new()}}{
Initialise Taxon LHT Grabber class
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TaxonPredictedLHTGetter$new(master_db, taxon)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{master_db}}{Fishlife database}

\item{\code{taxon}}{string holding the taxon name}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TaxonPredictedLHTGetter-extract"></a>}}
\if{latex}{\out{\hypertarget{method-TaxonPredictedLHTGetter-extract}{}}}
\subsection{Method \code{extract()}}{
Fetch the sought taxon details
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TaxonPredictedLHTGetter$extract()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TaxonPredictedLHTGetter-clone"></a>}}
\if{latex}{\out{\hypertarget{method-TaxonPredictedLHTGetter-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TaxonPredictedLHTGetter$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
