% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/taxon_predictor.R
\name{TaxonPredictor}
\alias{TaxonPredictor}
\title{Generate the new LHT matrix in shape and mathematical domain expected by Fishlife
Generate a matrix with the new predicted LHTs given some initial values, both in log and log-converted space
Build the new LHT matrix in the expected shape}
\description{
Generate the new LHT matrix in shape and mathematical domain expected by Fishlife
Generate a matrix with the new predicted LHTs given some initial values, both in log and log-converted space
Build the new LHT matrix in the expected shape

Generate the new LHT matrix in shape and mathematical domain expected by Fishlife
Generate a matrix with the new predicted LHTs given some initial values, both in log and log-converted space
Build the new LHT matrix in the expected shape
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-TaxonPredictor-new}{\code{TaxonPredictor$new()}}
\item \href{#method-TaxonPredictor-generate_new_lht_matrix}{\code{TaxonPredictor$generate_new_lht_matrix()}}
\item \href{#method-TaxonPredictor-predict}{\code{TaxonPredictor$predict()}}
\item \href{#method-TaxonPredictor-clone}{\code{TaxonPredictor$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TaxonPredictor-new"></a>}}
\if{latex}{\out{\hypertarget{method-TaxonPredictor-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TaxonPredictor$new(
  master_db,
  estimated_lhts,
  estimated_lht_cov,
  new_lhts,
  func_domains
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TaxonPredictor-generate_new_lht_matrix"></a>}}
\if{latex}{\out{\hypertarget{method-TaxonPredictor-generate_new_lht_matrix}{}}}
\subsection{Method \code{generate_new_lht_matrix()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TaxonPredictor$generate_new_lht_matrix()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TaxonPredictor-predict"></a>}}
\if{latex}{\out{\hypertarget{method-TaxonPredictor-predict}{}}}
\subsection{Method \code{predict()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TaxonPredictor$predict()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TaxonPredictor-clone"></a>}}
\if{latex}{\out{\hypertarget{method-TaxonPredictor-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TaxonPredictor$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
