% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/taxon_result_extractor.R
\name{TaxonResultExtractor}
\alias{TaxonResultExtractor}
\title{Add a prefix to the user-defined predicted LHT names
Fetch the sought LHT results and backtransform them as given by their corresponding conversion function}
\description{
Add a prefix to the user-defined predicted LHT names
Fetch the sought LHT results and backtransform them as given by their corresponding conversion function

Add a prefix to the user-defined predicted LHT names
Fetch the sought LHT results and backtransform them as given by their corresponding conversion function
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-TaxonResultExtractor-new}{\code{TaxonResultExtractor$new()}}
\item \href{#method-TaxonResultExtractor-extract}{\code{TaxonResultExtractor$extract()}}
\item \href{#method-TaxonResultExtractor-clone}{\code{TaxonResultExtractor$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TaxonResultExtractor-new"></a>}}
\if{latex}{\out{\hypertarget{method-TaxonResultExtractor-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TaxonResultExtractor$new(
  update_prefix,
  result_names,
  back_transform_matrix,
  results
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TaxonResultExtractor-extract"></a>}}
\if{latex}{\out{\hypertarget{method-TaxonResultExtractor-extract}{}}}
\subsection{Method \code{extract()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TaxonResultExtractor$extract()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TaxonResultExtractor-clone"></a>}}
\if{latex}{\out{\hypertarget{method-TaxonResultExtractor-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TaxonResultExtractor$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
