% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/updated_lht_generator.R
\name{UpdatedLHTGenerator}
\alias{UpdatedLHTGenerator}
\title{Multiple Updated LHT Generator class}
\value{
list of LHT matrices, where each name in the list corresponds a single taxon

a unnamed list of LHT values per taxon

renamed list of lht which names conform to Fishlife's expected patterns

list of LHT values (another list) for each taxon
}
\description{
Class that allows the streamline-generation of predicted LHTs given new data.
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-UpdatedLHTGenerator-new}{\code{UpdatedLHTGenerator$new()}}
\item \href{#method-UpdatedLHTGenerator-update}{\code{UpdatedLHTGenerator$update()}}
\item \href{#method-UpdatedLHTGenerator-clone}{\code{UpdatedLHTGenerator$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-UpdatedLHTGenerator-new"></a>}}
\if{latex}{\out{\hypertarget{method-UpdatedLHTGenerator-new}{}}}
\subsection{Method \code{new()}}{
Initialise the Updated LHT Generator class
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{UpdatedLHTGenerator$new(master_db, predicting_lht_df, func_domains, lht_names)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{master_db}}{Fishlife database}

\item{\code{predicting_lht_df}}{dataframe providing the taxon details and their predicting LHT values}

\item{\code{func_domains}}{list of transforming function which names must conform to FishLife's expectations}

\item{\code{lht_names}}{list of user-defined LHT names associated with their FishLife's counterparts}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-UpdatedLHTGenerator-update"></a>}}
\if{latex}{\out{\hypertarget{method-UpdatedLHTGenerator-update}{}}}
\subsection{Method \code{update()}}{
Update the given LHTs per taxon according to the new predicted value yielded by FishLfe
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{UpdatedLHTGenerator$update()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-UpdatedLHTGenerator-clone"></a>}}
\if{latex}{\out{\hypertarget{method-UpdatedLHTGenerator-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{UpdatedLHTGenerator$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
