% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/updated_lht_picker.R
\name{UpdatedLHTPicker}
\alias{UpdatedLHTPicker}
\title{Taxa's updated LHT picker}
\value{
a dataframe with the predicting and updated LHTs. The latter are prefixed with a keyword.
}
\description{
Fetch, transform and build a dataframe with the updated LHT values per taxon obtained from FishLife
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-UpdatedLHTPicker-new}{\code{UpdatedLHTPicker$new()}}
\item \href{#method-UpdatedLHTPicker-pick_and_backtransform}{\code{UpdatedLHTPicker$pick_and_backtransform()}}
\item \href{#method-UpdatedLHTPicker-clone}{\code{UpdatedLHTPicker$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-UpdatedLHTPicker-new"></a>}}
\if{latex}{\out{\hypertarget{method-UpdatedLHTPicker-new}{}}}
\subsection{Method \code{new()}}{
Initialise the UpdatedLHTPicker class
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{UpdatedLHTPicker$new(
  master_db,
  predicting_lht_df,
  updated_prefix,
  transform_function_list,
  backtransform_function_list,
  lht_names = NULL
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{master_db}}{Fishlife database}

\item{\code{predicting_lht_df}}{dataframe keeping the original inputted LHTs per taxon}

\item{\code{transform_function_list}}{list of transforming function which names must conform to FishLife's expectations}

\item{\code{backtransform_function_list}}{list of backward-transformation functions to be applied on obtained LHT from FishLife}

\item{\code{lht_names}}{list of user-defined LHT names associated with their FishLife's counterparts}

\item{\code{update_prefix}}{string to be added as prefix to the column names keeping the obtained new LHT values}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-UpdatedLHTPicker-pick_and_backtransform"></a>}}
\if{latex}{\out{\hypertarget{method-UpdatedLHTPicker-pick_and_backtransform}{}}}
\subsection{Method \code{pick_and_backtransform()}}{
Build a dataframe with the updated LHT for each wanted taxon as
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{UpdatedLHTPicker$pick_and_backtransform()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-UpdatedLHTPicker-clone"></a>}}
\if{latex}{\out{\hypertarget{method-UpdatedLHTPicker-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{UpdatedLHTPicker$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
